LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-gap-animation-002-expected.txt


FAIL column-gap: normal is not interpolable assert_equals: expected "100px" but got "80.24139404296875px"
